Added CI/CD workflows for automated testing and deployment  (#5)
* Update PR with automated changes
 (#3)

* add github actions

* print response

* set environment

* try fix

* add max tokens

* fix claude

* fix claude

* fix claude

* generate description

* fix

* revert

* fix

* fix

* fix

* fix

* fix

* fix

* Revert "fix"

This reverts commit 442b4a21efc968bacf0e03b5691edaa795c30015.

* overwrite description

* use claude for summary

* fix

* fix

* fix

* update title

* use data values

* try fix version check

* try fix version check 2

* fix test and bump version

* try fix title

* add nac
