+++ CONTRIBUTING.md
@@ -1,0 +1,15 @@
+# Contribution Guidelines
+
+## Pull Request Rules
+1. Each PR must include:
+   - Description of changes
+   - Related issue number (if applicable)
+   - Updated tests (if changing functionality)
+   - Updated version number in GmoCoinNet.csproj
+   - Updated documentation (if applicable)
+
+## Code Style
+1. Use C# naming conventions
+2. Include XML documentation for public APIs
+3. Keep methods focused and under 50 lines where possible
+4. Include unit tests for new functionality
